Fixed validation error...I believe...
I can write a test for this, but the basic issue is that when using `EnumRule` to validate, it calls `asEnum`, which attempts to create a new enum with the given input.  If this fails an exception is thrown...**or an instance of `TypeError`**.

It's the latter scenario that I found to be causing a problem.  I've simply added `Error` to the `catch` clause, which fixes it in my own tests.

Please let me know if you have any questions or concerns.

PS I know this is Hacktoberfest time and there are complaints about people spamming for t-shirts.  I'm a fan of Hacktoberfest and love the t-shirts, but this isn't spam.  I legit ran into the problem today and the fix **appears to be** quite simple.  Please give me the benefit of the doubt and let me know if you're otherwise worried.  Thx. 🤓